requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
python-dotenv>=1.0.0
loguru>=0.7.0
psutil>=5.9.0
anthropic>=0.8.0
pillow>=10.0.0
numpy>=1.24.0 